add webauthn route
