+++ internal/httpserve/route/router.go
@@ -205,6 +205,7 @@ func RegisterRoutes(router *Router) error {
 		registerWebhookHandler,
 		register2faHandler,
 		registerExampleCSVHandler,
+		registerWebAuthnWellKnownHandler,
 	}
 
 	if router.LocalFilePath != "" {
