+++ internal/httpserve/route/webauthn
@@ -1,0 +1,8 @@
+{
+    "origins": [
+        "https://www.theopenlane.io",
+        "https://theopenlane.io",
+        "https://console.theopenlane.io",
+        "https://api.theopenlane.io"
+    ]
+}
